ai_daily_report/requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dotenv>=1.0.0
anthropic>=0.18.0
